Moving logError from cli to usage-data
