+++ packages/custom-functions-metadata/package.json
@@ -28,7 +28,6 @@
     "assert": "^1.5.0",
     "commander": "^6.2.0",
     "nconf": "^0.11.3",
-    "office-addin-cli": "^1.3.4",
     "xregexp": "^4.3.0"
   },
   "devDependencies": {
@@ -41,6 +40,7 @@
     "mocha": "^9.1.1",
     "office-addin-lint": "^1.4.5",
     "office-addin-prettier-config": "^1.1.3",
+    "office-addin-usage-data": "^1.4.5",
     "rimraf": "^3.0.2",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.5"
