+++ packages/office-addin-cli/src/main.ts
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-export * from "./log";
 export * from "./npmPackage";
 export * from "./parse";
