+++ packages/office-addin-dev-certs/src/commands.ts
@@ -2,7 +2,8 @@
 // Licensed under the MIT license.
 
 import * as commander from "commander";
-import { logErrorMessage, parseNumber } from "office-addin-cli";
+import { parseNumber } from "office-addin-cli";
+import { logErrorMessage } from "office-addin-usage-data";
 import * as defaults from "./defaults";
 import { ensureCertificatesAreInstalled } from "./install";
 import { deleteCertificateFiles, uninstallCaCertificate } from "./uninstall";
