+++ packages/office-addin-dev-settings/package.json
@@ -28,7 +28,6 @@
     "inquirer": "^7.3.3",
     "jszip": "^3.7.1",
     "junk": "^3.1.0",
-    "office-addin-cli": "^1.3.4",
     "office-addin-manifest": "^1.7.6",
     "office-addin-usage-data": "^1.4.5",
     "open": "^6.4.0",
