+++ packages/office-addin-lint/package.json
@@ -31,7 +31,6 @@
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-office-addins": "^1.1.4",
     "eslint-plugin-prettier": "^3.3.1",
-    "office-addin-cli": "^1.3.4",
     "office-addin-prettier-config": "^1.1.3",
     "office-addin-usage-data": "^1.4.5",
     "prettier": "^2.2.1"
