+++ packages/office-addin-manifest/package.json
@@ -26,7 +26,6 @@
     "commander": "^6.2.0",
     "jszip": "^3.7.1",
     "node-fetch": "^2.6.1",
-    "office-addin-cli": "^1.3.4",
     "office-addin-usage-data": "^1.4.5",
     "path": "^0.12.7",
     "uuid": "^8.3.2",
