+++ packages/office-addin-test-server/src/cli.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import * as commander from "commander";
-import { logErrorMessage } from "office-addin-cli";
+import { logErrorMessage } from "office-addin-usage-data";
 import * as commands from "./commands";
 import { defaultPort } from "./testServer";
