+++ packages/office-addin-usage-data/package.json
@@ -23,7 +23,6 @@
   "dependencies": {
     "applicationinsights": "^1.7.3",
     "commander": "^6.2.0",
-    "office-addin-cli": "^1.3.4",
     "readline-sync": "^1.4.9"
   },
   "devDependencies": {
@@ -37,6 +36,7 @@
     "mocha": "^9.1.1",
     "office-addin-prettier-config": "^1.1.3",
     "rimraf": "^3.0.2",
+    "sinon": "^7.5.0",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.5"
   },
