+++ packages/office-addin-usage-data/src/cli.ts
@@ -4,7 +4,7 @@
 // Licensed under the MIT license.
 
 import * as commander from "commander";
-import { logErrorMessage } from "office-addin-cli";
+import { logErrorMessage } from "./log";
 import * as commands from "./command";
 
 /* global process */
