+++ packages/office-addin-usage-data/src/log.ts
