+++ packages/office-addin-usage-data/src/main.ts
@@ -2,5 +2,6 @@
 // Licensed under the MIT license.
 
 export * from "./defaults";
+export * from "./log";
 export * from "./usageData";
 export * from "./usageDataSettings";
